x64 works (32-bit dropped for now)
I did not realize we can do it in x64. The remaining piece of the puzzle
was restoring the FS/GS base which gets overwritten by FS/GS selector loads.

+++ src/msg/dump.cpp
@@ -67,7 +67,7 @@ void QnxMsg::dump_message(FILE* s, const QnxMessageList& list,  Msg& msg) {
     fprintf(s, "message %s {\n", t->m_name);
     for(size_t fi = 0; fi < t->m_field_count; ++fi) {
         auto& f = t->m_fields[fi];
-        fprintf(s, "@%02x ",  f.m_offset);
+        fprintf(s, "@%02x ", static_cast<uint32_t>(f.m_offset));
         fprintf(s, "   %s: ", f.m_name);
 
         auto field_data = static_cast<void*>(msg_buf.get() + f.m_offset);
